+++ Private/Connect-SiteServer.ps1
@@ -1,7 +1,7 @@
 <#
 .Synopsis
 Created on:   26/10/2023
-Updated on:   20/01/2024
+Updated on:   17/02/2024
 Created by:   Ben Whitmore
 Filename:     Connect-SiteServer.ps1
 
@@ -52,7 +52,7 @@ function Connect-SiteServer {
             Write-Log -Message "Warning: Could not import the ConfigurationManager.psd1 Module"
             Write-Warning "Warning: Could not import the 'ConfigurationManager.psd1' module"
             Write-Log -Message ("'{0}'" -f $_.Exception.Message) -LogId $LogId -Severity 3
-            Get-ScriptEnd -LogId $LogId -Message $_.Exception.Message
+            Get-ScriptEnd -LogId $LogId -ErrorMessage $_.Exception.Message
     }
 
         # Check the SMS Provider is valid
@@ -60,7 +60,7 @@ function Connect-SiteServer {
             Write-Log -Message ("Could not connect to the Provider '{0}'" -f $ProviderMachineName) -Severity 3
             Write-Warning ("Could not connect to the Provider '{0}' `nDid you specify the correct Site System?" -f $ProviderMachineName)
             Write-Log -Message ("'{0}'" -f $_.Exception.Message) -LogId $LogId -Severity 3
-            Get-ScriptEnd -LogId $LogId -Message $_.Exception.Message
+            Get-ScriptEnd -LogId $LogId -ErrorMessage $_.Exception.Message
         }
         else {
             Write-Log -Message ("Connected to provider {0} at site '{1}'" -f $ProviderMachineName, $SiteCode )
@@ -72,7 +72,7 @@ function Connect-SiteServer {
             if (!($SiteCode -eq ( Get-PSDrive -ErrorAction SilentlyContinue | Where-Object { $_.Provider -like "*CMSite*" }).Name) ) {
                 Write-Log -Message ("No PSDrive found for '{0}' in PSProvider CMSite for Root '{1}'" -f $SiteCode, $ProviderMachineName) -LogId $LogId -Severity 3
                 Write-Warning ("No PSDrive found for '{0}' in PSProvider CMSite for Root '{1}'. Did you specify the correct Site Code?" -f $SiteCode, $ProviderMachineName)
-                Get-ScriptEnd -LogId $LogId -Message $_.Exception.Message
+                Get-ScriptEnd -LogId $LogId -ErrorMessage $_.Exception.Message
 
             }
             else {
@@ -84,7 +84,7 @@ function Connect-SiteServer {
         catch {
             Write-Log -Message ("Warning: Could not connect to the specified provider '{0}' at site '{1}'" -f $ProviderMachineName, $SiteCode) -LogId $LogId -Severity 3
             Write-Warning ("Warning: Could not connect to the specified provider '{0}' at site '{1}'" -f $ProviderMachineName, $SiteCode)
-            Get-ScriptEnd -Message $_.Exception.Message -LogId $LogId 
+            Get-ScriptEnd -ErrorMessage $_.Exception.Message -LogId $LogId 
         }
     }
 }